Adjust on delete method to complain with swagger description

+++ src/main/java/net/superbid/java/backend/service/BlogService.java
@@ -29,7 +29,8 @@ public class BlogService {
 	}
 	
 	public void deletePost(Long id) { 
-		postRepository.delete(new Post(id , null , null , null));
+		if (postRepository.existsById(id))
+			postRepository.delete(new Post(id , null , null , null));
 	}
 	
 	public Message findPost(Long id) {
